dashboard/requirements.txt
fastapi
uvicorn[standard]
pyserial
scikit-learn
numpy
joblib
jinja2
python-multipart
